Create String Game
Problem
Alice and Bob are playing a game. They have a common string SS of length NN. The players also have their individual strings AA (belonging to Alice) and BB (belonging to Bob) which are empty in the beginning. Game begins with Alice and both players take alternate turns.

In her/his turn, the player picks a single character from string SS, adds it to the end of their individual string and deletes the picked character from string SS.

The game continues until string SS is empty. Find whether there exists a sequence of moves such that the strings AA and BB are same at the end of the game.

Input Format
The first line of input will contain a single integer TT, denoting the number of test cases.
Each test case consists of multiple lines of input.
The first line of each test case contains an integer NN — the length of the string.
The next line contains the strings SS consisting of lowercase english alphabets.
Output Format
For each test case, output on a new line, YES if there exists a sequence of moves such that the strings AA and BB are same at the end of the game, and NO otherwise.

You may print each character of the string in uppercase or lowercase (for example, the strings YES, yEs, yes, and yeS will all be treated as identical).

Constraints
1 \leq T \leq 10^31≤T≤10 
3
 
1 \leq N \leq 10^51≤N≤10 
5
 
SS consists of lowercase english alphabets
The sum of NN over all test cases does not exceed 2\cdot 10^52⋅10 
5
 .
Sample 1:
Input
Output
4
4
abab
5
cbcba
4
abcd
6
pqprqr
YES
NO
NO
YES
Explanation:
Test case 11: Consider the following sequence of moves:

Alice picks the first character of string SS and adds it to the end of string AA. Thus, SS becomes bab and AA becomes a.
Bob picks the second character of string SS and adds it to the end of string BB. Thus, the strings are S =S= bb, A =A= a, and B =B= a .
Alice picks the second character of string SS and adds it to the end of string AA. Thus, the strings are S =S= b, A =A= ab, and B =B= a .
Bob picks the first character of string SS and adds it to the end of string BB. Thus, SS becomes empty, A =A= ab, and B =B= ab .
We can see that using this sequence of moves, the final strings AA and BB are equal.

Test case 22: There exists no sequence of moves such that the strings AA and BB become equal in the end.

Test case 33: There exists no sequence of moves such that the strings AA and BB become equal in the end.

Test case 44: Consider the following sequence of moves:

Alice picks the first character of string SS and adds it to the end of string AA. Thus, SS becomes qprqr and AA becomes p.
Bob picks the second character of string SS and adds it to the end of string BB. Thus, the strings are S =S= qrqr, A =A= p, and B =B= p .
Alice picks the second character of string SS and adds it to the end of string AA. Thus, the strings are S =S= qqr, A =A= pr, and B =B= p .
Bob picks the third character of string SS and adds it to the end of string BB. Thus, SS becomes qq, AA becomes pr, and BB becomes pr.
Alice picks the second character of string SS and adds it to the end of string AA. Thus, the strings are S =S= q, A =A= prq, and B =B= pr .
Bob picks the first character of string SS and adds it to the end of string BB. Thus, SS becomes empty, A =A= prq, and B =B= prq .
We can see that using this sequence of moves, the final strings AA and BB are equal.