Create MaxEdges
Problem
Tracy gives Charlie a Directed Acyclic Graph with NN vertices. Among these NN vertices, KK vertices are sources, and LL vertices are sinks.

Find out the maximum number of edges this graph can have.

Note:

A source is a vertex with no incoming edge.
A sink is a vertex with no outgoing edge.
A vertex can be both, a source, and a sink.
Input Format
First line will contain TT, number of test cases. Then the test cases follow.
Each test case contains of a single line of input, three space-separated integers N, K,N,K, and LL - the number of vertices, the number of sources and the number of sinks respectively.
Output Format
For each test case, output in a single line, the maximum number of edges this graph can have.
Constraints
1 \leq T \leq 10^51≤T≤10 
5
 
2 \leq N \leq 10^92≤N≤10 
9
 
1 \leq K \lt N1≤K<N
1 \leq L \lt N1≤L<N
Sample 1:
Input
Output
2
3 1 1
5 3 3
3
4
Explanation:
Test case 11: Assume that the vertices are numbered 1,2,1,2, and 33. Let 11 be a source and 33 be a sink. The edges of a possible DAG are 1\rightarrow 2, 2\rightarrow 3,1→2,2→3, and 1\rightarrow 31→3.

The number of edges in this graph are 33. It can be shown that this is the maximum number of edges possible under given constraints.

Test case 22: Assume that the vertices are numbered 1,2,3,4,1,2,3,4, and 55. Let 1,2,1,2, and 33 be sources and 3, 4,3,4, and 55 be sinks. The edges of a possible DAG are 1\rightarrow 4, 2\rightarrow 4, 1\rightarrow 51→4,2→4,1→5 and 2\rightarrow 52→5.

The number of edges in this graph are 44. It can be shown that this is the maximum number of edges possible under given constraints.