Create  Thermostat
Vlad came home and found out that someone had reconfigured the old thermostat to the temperature of a.

The thermostat can only be set to a temperature from l to r inclusive, the temperature cannot change by less than x. Formally, in one operation you can reconfigure the thermostat from temperature a to temperature b if |a−b|≥x and l≤b≤r.

You are given l, r, x, a and b. Find the minimum number of operations required to get temperature b from temperature a, or say that it is impossible.

Input
The first line of input data contains the single integer t (1≤t≤104) — the number of test cases in the test.

The descriptions of the test cases follow.

The first line of each case contains three integers l, r and x (−109≤l≤r≤109, 1≤x≤109) — range of temperature and minimum temperature change.

The second line of each case contains two integers a and b (l≤a,b≤r) — the initial and final temperatures.

Output
Output t numbers, each of which is the answer to the corresponding test case. If it is impossible to achieve the temperature b, output -1, otherwise output the minimum number of operations.