Create Minimum Number of Ones
Problem
Your teacher gave you an assignment — given an integer NN, construct a binary string B = b_1b_2b_3\ldots b_NB=b 
1
​
 b 
2
​
 b 
3
​
 …b 
N
​
  of length NN such that:

\max(b_i, b_{i+1}) = 1
max(b 
i
​
 ,b 
i+1
​
 )=1
for every ii from 11 to N-1N−1.

What is the minimum number of 11's such a binary string can contain?

Note: A binary string is a string consisting of only the digits 00 and 11.

Input Format
The first line of input will contain a single integer TT, denoting the number of test cases.
Each test case contains a single integer NN — the length of binary string you'd like to construct.
Output Format
For each test case, output on a new line the minimum number of 11's required to complete the assignment.

Constraints
1 \leq T \leq 10001≤T≤1000
2 \leq N \leq 10002≤N≤1000
Sample 1:
Input
Output
6
6
8
2
3
5
100
3
4
1
1
2
50
Explanation:
Test case 11: One possible binary string is \texttt{010101}010101. This has three 11's, and it can be verified that the maximum of any two adjacent characters is 11.
Achieving this with less than three 11's is not possible.

Test case 22: One possible binary string is \texttt{10101010}10101010. This has four 11's, and it can be verified that the maximum of any two adjacent characters is 11.
Achieving this with less than four 11's is not possible.

Test case 33: One possible binary string is \texttt{10}10. This has one 11, and it can be verified that the maximum of any two adjacent characters is 11.

Test case 44: One possible binary string is \texttt{010}010. This has one 11, and it can be verified that the maximum of any two adjacent characters is 11.