Create Expensive Steps
Problem
There is an N\times NN×N grid, with rows numbered 11 to NN from top to bottom and columns numbered 11 to NN from left to right. The cell at the intersection of the ii-th row and jj-th column is denoted (i, j)(i,j).

Chef is standing at position (x_1, y_1)(x 
1
​
 ,y 
1
​
 ) of this grid. He wants to reach position (x_2, y_2)(x 
2
​
 ,y 
2
​
 ).

To achieve this, he can take either horizontal or vertical steps. That is, from cell (x, y)(x,y), he can move to one of (x-1, y), (x+1, y), (x, y-1), (x, y+1)(x−1,y),(x+1,y),(x,y−1),(x,y+1).

However, Chef does not need to stay within the grid, that is, he can choose to go outside it if he so wishes.

The cost of a horizontal/vertical step is 00 if both the starting and ending square lie outside the grid, and 11 otherwise.
Formally, suppose Chef takes one step from (x, y)(x,y) to (x', y')(x 
′
 ,y 
′
 ). Then,

If (x, y)(x,y) lies within the grid, the cost of this step is 11
If (x', y')(x 
′
 ,y 
′
 ) lies within the grid, the cost of this step is 11
If both the above don't hold, the cost of this step is 00
Find the minimum cost to reach (x_2, y_2)(x 
2
​
 ,y 
2
​
 ) from (x_1, y_1)(x 
1
​
 ,y 
1
​
 ).

Input Format
The first line of input will contain a single integer TT, denoting the number of test cases.
Each test case consists of one line of input containing five space-separated integers NN, x_1x 
1
​
 , y_1y 
1
​
 , x_2x 
2
​
 , and y_2y 
2
​
 .
Output Format
For each test case, output on a new line the minimum cost of traveling from (x_1, y_1)(x 
1
​
 ,y 
1
​
 ) to (x_2, y_2)(x 
2
​
 ,y 
2
​
 ).

Constraints
1 \leq T \leq 10^{5}1≤T≤10 
5
 
1 \leq N \leq 10^{5}1≤N≤10 
5
 
1 \leq x_1,y_1,x_2,y_2 \leq N1≤x 
1
​
 ,y 
1
​
 ,x 
2
​
 ,y 
2
​
 ≤N
Sample 1:
Input
Output
6
3 1 1 2 3
5 2 3 3 4
3 1 1 2 1
6 2 1 5 5
8 3 4 6 7
10 4 5 9 9
2
2
1
3
5
6
Explanation:
Test case 11: One optimal sequence of moves is as follows:

Chef moves from (1, 1)(1,1) to (0, 1)(0,1) with a cost of 11. Note that (0, 1)(0,1) is outside the grid.
Now, Chef can move on the outside from (0, 1)(0,1) to (2, 4)(2,4) with no cost.
Finally, Chef moves from (2, 4)(2,4) to (2, 3)(2,3) with cost 11, for a total cost of 22.
Test case 22: Chef can move as follows: (2, 3) \to (3, 3) \to (3, 4)(2,3)→(3,3)→(3,4), for a cost of 22.

Test case 44: Chef can move as follows: (2, 1) \to (2, 0)(2,1)→(2,0) with cost 11, then (2, 0) \to (5, 7)(2,0)→(5,7) on the outside with cost 00, then (5, 7) \to (5, 6) \to (5, 5)(5,7)→(5,6)→(5,5) with cost 22 for a total cost of 33